requirements.txt
Flask
Flask-CORS
word2number
requests
ratelimit
python-dotenv
openalgo